Move modification of constructor super calls to PotentialMethod. Had to give PotentialMethod visibility of PotentialClass so it is  able to determine what the super class is, if anything at all.
